convert INSERT to a listener, start UPDATE

+++ public/index.js
@@ -2,7 +2,10 @@
 // globals
 // **********************************************
 const workoutListElem = $("#workoutList"); 
-//const workoutDeleteButton = $(".deleteButton"); 
+const submitBtnElem = $("#submitBtn"); 
+var workoutNameElem = $("#workoutName"); 
+var caloriesBurnedElem = $("#caloriesBurned"); 
+var workoutID; 
 
 // **********************************************
 // functions 
@@ -28,12 +31,14 @@ function loadWorkouts (data) {
     <div class="card">
       <h5 class="card-title">You worked out ${data[i].workoutCreated}</h5>
       <div class="card-body"></div>
-        <p class="card-text">${data[i].workoutName}</p>
-        <p class="card-text">${data[i].caloriesBurned} calories </p>
+        <p data-id="${data[i]._id}" class="card-text workoutName">${data[i].workoutName}</p>
+        <p data-id="${data[i]._id}" class="card-text caloriesBurned">${data[i].caloriesBurned} calories </p>
+        <a href="#" class="btn btn-primary updateButton" data-id="${data[i]._id}">Update</a>
         <a href="#" class="btn btn-primary deleteButton" data-id="${data[i]._id}">Delete</a>
       </div>
     </div>
-    <br>`
+    <br>
+    <hr>`
 
     workoutListElem.append (myCard); 
 
@@ -71,12 +76,17 @@ function init () {
 // listeners 
 // **********************************************
 
+// **********************************************
+// delete 
+// **********************************************
+
 workoutListElem.on("click", function(e) {
   event.preventDefault(); 
   element = e.target;
   data_id = element.getAttribute("data-id");
-  console.log ("delete id " + data_id); 
-  if (data_id){
+
+  if (data_id && element.classList.contains("deleteButton")) {
+    console.log ("delete id " + data_id); 
     fetch("/delete/" + data_id, {
       method: "delete"
     })
@@ -96,7 +106,52 @@ workoutListElem.on("click", function(e) {
       console.log("Fetch Error :-S", err);
     });
   }
+
+  if (data_id && element.classList.contains("updateButton")) {
+
+    var workoutName;
+    var caloriesBurned; 
+
+    console.log ("update " + data_id); 
+
+    myCardTextList = document.querySelectorAll (".card-text.workoutName"); 
+    for (let i = 0;i<myCardTextList.length;i++){
+      if (myCardTextList[i].getAttribute("data-id") == data_id){
+        workoutName = myCardTextList[i].innerHTML; 
+      }
+    }
+    myCardTextList = document.querySelectorAll (".card-text.caloriesBurned"); 
+    for (let i = 0;i<myCardTextList.length;i++){
+      if (myCardTextList[i].getAttribute("data-id") == data_id){
+        caloriesBurned = myCardTextList[i].innerHTML; 
+      }
+    }
+    console.log ("w " + workoutName + " cb " + caloriesBurned); 
+    workoutNameElem.val (workoutName); 
+    caloriesBurnedElem.val (caloriesBurned); 
+
+  }
+
+
 }); // workoutListElem click 
+
+// **********************************************
+// submit (INSERT or UPDATE)
+// **********************************************
+
+submitBtnElem.on("click", function(e) {
+  event.preventDefault(); 
+  element = e.target;
+  console.log ("insert"); 
+
+  var newWorkout = {
+    workoutName: workoutNameElem.val().trim(),
+    caloriesBurned: caloriesBurnedElem.val().trim()}; 
+  $.post("/submit", newWorkout).then(function(dbWorkout){
+    console.log("insert ok" + dbWorkout);
+    location.reload();
+  });
+}); // submitBtnElem click 
 
 // **********************************************
 // init
